% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm_multi}
\alias{lm_multi}
\title{Linear Modeling of Multivariate Response}
\usage{
lm_multi()
}
\value{
A list with the training function and the prediction function.
}
\description{
This model is fed to \code{\link{conformal.multidim.full}},
\code{\link{conformal.multidim.split}}, and \code{\link{conformal.multidim.msplit}}.
It outputs a training function and a prediction function.
}
\details{
The training function takes as input:

x The feature matrix  (dim n x p)
y The response matrix (dim n x q)

The predict function, instead, takes as input:

out The output of a previous call to train.fun
newx The new features to evaluate (i.e. an n0 x p matrix)
Here I defined an lm model for every dimension of the responses (q).
}
