% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_multi}
\alias{mean_multi}
\title{Mean of Multivariate Response}
\usage{
mean_multi()
}
\value{
A list with two components:
\item{train.fun}{Function to train the model.}
\item{predict.fun}{Function to make predictions on new data.}
}
\description{
This model can be used with conformal prediction functions.
It returns a training function and a prediction function.
}
\details{
The training function takes as input:
\describe{
  \item{x}{Feature matrix of dimension n x p.}
  \item{y}{Response matrix of dimension n x q.}
}

The prediction function takes as input:
\describe{
  \item{out}{Output of a previous call to \code{train.fun}.}
  \item{newx}{New feature matrix to evaluate, dimension n0 x p.}
}
}
\seealso{
\code{\link{conformal.multidim.split}}
}
