% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_regression.R
\name{computing_s_regression}
\alias{computing_s_regression}
\title{Compute modulation function for residuals}
\usage{
computing_s_regression(mat_residual, type, alpha, tau)
}
\arguments{
\item{mat_residual}{A vector of residuals obtained via multivariate modeling.}

\item{type}{A string indicating the type of modulation function. Options are
"identity", "st-dev", or "alpha-max".}

\item{alpha}{The confidence level for the interval.}

\item{tau}{A number between 0 and 1 used for the randomized version of the algorithm.}
}
\value{
Local scoring values for the residuals.
}
\description{
Helper function used internally by \code{\link{conformal.multidim.split}} and
\code{\link{conformal.multidim.msplit}}.
}
\references{
Diquigiovanni, Fontana, Vantini (2021), "Conformal Prediction Bands for Multivariate Functional Data"
}
\seealso{
\code{\link{conformal.multidim.split}}, \code{\link{conformal.multidim.msplit}}
}
